+++ src/Commands/ToDo/NewTodoTaskChecklistItem.cs
@@ -1,18 +1,22 @@
-﻿using PnP.PowerShell.Commands.Attributes;
+using PnP.PowerShell.Commands.Attributes;
 using PnP.PowerShell.Commands.Base;
+using PnP.PowerShell.Commands.Base.Completers;
 using PnP.PowerShell.Commands.Base.PipeBinds;
 using PnP.PowerShell.Commands.Utilities;
 using System.Management.Automation;
 
 namespace PnP.PowerShell.Commands.ToDo
 {
-    [Cmdlet(VerbsData.Update, "PnPTodoList")]
+    [Cmdlet(VerbsCommon.New, "PnPTodoTaskChecklistItem")]
     [RequiredApiDelegatedOrApplicationPermissions("graph/Tasks.ReadWrite")]
     [ApiNotAvailableUnderApplicationPermissions]
-    public class UpdateTodoList : PnPGraphCmdlet
+    public class NewTodoTaskChecklistItem : PnPGraphCmdlet
     {
+        [Parameter(Mandatory = true), ArgumentCompleter(typeof(TodoListCompleter))]
+        public string List;
+
         [Parameter(Mandatory = true)]
-        public string Identity;
+        public TodoTaskPipeBind Task;
 
         [Parameter(Mandatory = true)]
         public string DisplayName;
@@ -28,8 +32,14 @@ namespace PnP.PowerShell.Commands.ToDo
                 return;
             }
 
-            var todoList = ToDoUtility.UpdateList(GraphRequestHelper, url, Identity, DisplayName);
-            WriteObject(todoList, false);
+            var listId = ToDoUtility.GetListId(GraphRequestHelper, url, List);
+            if (listId == null)
+            {
+                throw new PSArgumentException("Todo list not found", nameof(List));
+            }
+
+            var checklistItem = ToDoUtility.CreateChecklistItem(GraphRequestHelper, url, listId, Task.Id, DisplayName);
+            WriteObject(checklistItem, false);
         }
     }
 }