+++ src/Commands/ToDo/RemoveTodoTaskLinkedResource.cs
@@ -1,0 +1,53 @@
+using PnP.PowerShell.Commands.Attributes;
+using PnP.PowerShell.Commands.Base;
+using PnP.PowerShell.Commands.Base.Completers;
+using PnP.PowerShell.Commands.Base.PipeBinds;
+using PnP.PowerShell.Commands.Utilities;
+using System.Management.Automation;
+
+namespace PnP.PowerShell.Commands.ToDo
+{
+    [Cmdlet(VerbsCommon.Remove, "PnPTodoTaskLinkedResource")]
+    [RequiredApiDelegatedOrApplicationPermissions("graph/Tasks.ReadWrite")]
+    [ApiNotAvailableUnderApplicationPermissions]
+    public class RemoveTodoTaskLinkedResource : PnPGraphCmdlet
+    {
+        [Parameter(Mandatory = true), ArgumentCompleter(typeof(TodoListCompleter))]
+        public string List;
+
+        [Parameter(Mandatory = true)]
+        public TodoTaskPipeBind Task;
+
+        [Parameter(Mandatory = true)]
+        public string Identity;
+
+        [Parameter(Mandatory = false)]
+        public EntraIDUserPipeBind User;
+
+        protected override void ExecuteCmdlet()
+        {
+            var url = ToDoUtility.GetTodoRootUrl(this, ParameterSpecified(nameof(User)) ? User : null);
+            if (url == null)
+            {
+                return;
+            }
+
+            var listId = ToDoUtility.GetListId(GraphRequestHelper, url, List);
+            if (listId == null)
+            {
+                throw new PSArgumentException("Todo list not found", nameof(List));
+            }
+
+            var graphResult = ToDoUtility.DeleteLinkedResource(GraphRequestHelper, url, listId, Task.Id, Identity);
+
+            if (graphResult.StatusCode == System.Net.HttpStatusCode.NoContent)
+            {
+                LogDebug("Todo task linked resource deleted successfully");
+            }
+            else
+            {
+                throw new PSArgumentException("Todo task linked resource could not be deleted", nameof(Identity));
+            }
+        }
+    }
+}